ceph-backport.sh: implement interactive setup routine
Instead of offering the user setup advice in the form of documentation,
provide an interactive setup routine that can be run any number of times.

This commit also deprecates the ~/bin/backport_common.sh file in favor
of a set of files:

    ~/.github_user
    ~/.github_token
    ~/.redmine_user_id
    ~/.redmine_key

(NOTE: The user need not care about these files. Their purpose is to save the
results of the interactive setup routine and make those available to future runs
of the script.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>