qa/deepsea: run the default validation tests only once
The Validation class in deepsea.py has a feature where certain tests are
configured to run every time the class/subtask is invoked. When the
class/subtask is invoked more than once, this is not a good thing because
the default validation tests get run on every invocation.

This patch fixes the issue by ensuring that the default tests are run only
once.

Signed-off-by: Nathan Cutler <ncutler@suse.com>