qa/rados/cephadm/smoke: restore opensuse_15.2
openSUSE 15.2 was in this suite for a short time, but was dropped when
it started to fail due to an issue with the testing infrastructure.

That issue having been resolved, and other blockers removed, we are now
re-enabling openSUSE 15.2 as a target distro for cephadm smoke-testing
in the Sepia lab.

Signed-off-by: Nathan Cutler <ncutler@suse.com>